git: credential-store
